+++ lib/tests/test_conflicts.rs
@@ -25,6 +25,7 @@ use jj_lib::conflicts::parse_conflict;
 use jj_lib::conflicts::update_from_content;
 use jj_lib::files::FileMergeHunkLevel;
 use jj_lib::merge::Merge;
+use jj_lib::merge::SameChange;
 use jj_lib::merged_tree::MergeOptions;
 use jj_lib::repo::Repo as _;
 use jj_lib::repo_path::RepoPath;
@@ -2390,6 +2391,7 @@ fn materialize_conflict_string(
         marker_len: None,
         merge: MergeOptions {
             hunk_level: FileMergeHunkLevel::Line,
+            same_change: SameChange::Accept,
         },
     };
     String::from_utf8(materialize_merge_result_to_bytes(&contents, &options).into()).unwrap()